Cleaned code
created classification sub folder and moved all .py files there
added few readme files
