+++ README.md
@@ -7,6 +7,12 @@ All work within Mimick and News-Media-Reliability folders are the explicit work 
 whose information can be found below.  These folders are submodules of original author's github
 repos.
 
+### News Media Reliability 
+Git Hub repo link for News Media Reliability code base is https://github.com/ramybaly/News-Media-Reliability
+
+### Mimick
+Git Hub repo link for Mimick code base is https://github.com/yuvalpinter/Mimick
+
 ## Student Work Overview
 
 This repo contains work related to project proposal which can be viewed via following link:
@@ -15,5 +21,5 @@ https://docs.google.com/document/d/1Vn2480AQH09SXASAashxWCgOlh11qdBcBBA5MiEzVcs/
 
 The final write-up for this project can be found here:
 
-<Placeholder>
+https://docs.google.com/document/d/18fOmQLKhdbwbJXZ9a96_sQXVhhvDuHBiekh_gH4_R5Y/edit?usp=sharing
