+++ original_work/README.md
@@ -1,8 +1,7 @@
 # Original Student Work for UC Berkeley W266
 
 
-## To recreate the embeddings from scratch, execute below command
-
-python wiki_parser.py
-
-## To run models on top of embeddings generated execute notebook results_comparisons
+## Accounting For Out Of Vocab Words In Fake News & Bias Detection
+This project is taken as part if W266 Final project by Clay Leach and Gaurav Desai
+All the original code and analysis done by us is stored in this 'original_work' folder
+Any code that we have reused from two reference papers is stored in their respective folder unless otherwise mentioned.
