+++ original_work/features/README.md
@@ -1,0 +1,2 @@
+### This directory stores numpy files (.npy) for each wiki feature per model.
+Since the feature file names are same for every model, we create separate sub folder with model's name and store .npy files generated by a particular model under that model's sub folder.
