+++ original_work/wikipedia_data/README.md
@@ -1,4 +1,6 @@
 # Wikipedia Data
 
-Folder to house Wikipedia text data for news sites which will be used in conjunction with Mimick models to 
-replace original wikipedia word2vec features in classification task
+Folder to house Wikipedia content data for news sites which will be used in conjunction with Mimick models to replace original wikipedia word2vec features in classification task
+Folder contains two files
+1. wiki_corpus.csv -> Contains wikipedia extracts after running wiki_parser.py in classification module
+2. wiki_manual_searched_pages.csv -> Contains website url and correspoding wikipedia page url for websites which do not get automatically mapped to their wikipedia page via wiki API
